Fixing

+++ server/src/server/server.c
@@ -670,7 +670,8 @@ int c_server_get_display_data(int display_id, void *display_data) {
 	*/
 
 	int i, flag;
-	float *data = (float*)display_data;
+	//float *data = (float*)display_data;
+	double *data = (double*)display_data;
 	flag = 0;
 	int pan_sz;
